+++ .changeset/tender-books-cheer.md
@@ -1,0 +1,5 @@
+---
+"@jpmorganchase/uitk-core": patch
+---
+
+PillBase moved to root folder of Pill; RadioButton subcomponents to begin with component name
