+++ packages/core/src/radio-button/RadioButton.tsx
@@ -2,8 +2,8 @@ import classnames from "classnames";
 import { ChangeEventHandler, FC, forwardRef, HTMLAttributes } from "react";
 import { ControlLabel, ControlLabelProps } from "../control-label";
 import { makePrefixer } from "../utils";
-import { Radio } from "./Radio";
-import { RadioIconProps } from "./RadioIcon";
+import { RadioButtonBase as Radio } from "./RadioButtonBase";
+import { RadioButtonIconProps } from "./RadioButtonIcon";
 
 import "./RadioButton.css";
 
@@ -17,7 +17,7 @@ export interface RadioButtonProps
   /**
    * Custom icon component
    */
-  icon?: FC<RadioIconProps>;
+  icon?: FC<RadioButtonIconProps>;
   /**
    * The label to be shown next to the radio
    */
