+++ packages/core/src/radio-button/RadioButtonBase.css
@@ -1,5 +1,5 @@
 /* Styles applied to root component */
-.uitkRadio {
+.uitkRadioButtonBase {
   color: var(--uitk-text-secondary-foreground);
   padding: 1px 0;
   flex: 0 0 auto;
@@ -22,7 +22,7 @@
 }
 
 /* Styles applied to container of Radio input and icon */
-.uitkRadio-radioContainer {
+.uitkRadioButtonBase-radioContainer {
   width: 100%;
   display: flex;
   align-items: inherit;
@@ -30,7 +30,7 @@
 }
 
 /* Styles applied to input component */
-.uitkRadio-input {
+.uitkRadioButtonBase-input {
   cursor: inherit;
   position: absolute;
   opacity: 0;
@@ -44,8 +44,8 @@
 }
 
 /* Styles applied to root component when state `focusVisible={true}` */
-.uitkRadio-focusVisible,
-.uitkRadio:focus {
+.uitkRadioButtonBase-focusVisible,
+.uitkRadioButtonBase:focus {
   outline: var(--uitk-focused-outline);
   outline-offset: var(--uitk-focused-outlineOffset);
 }
