+++ packages/core/src/radio-button/RadioButtonGroup.tsx
@@ -8,13 +8,13 @@ import {
   useRef,
   useState,
 } from "react";
-import { FormLabel } from "../form-field";
 import { useFormFieldProps } from "../form-field-context";
 import { FormGroup } from "../form-group";
+import { FormLabel } from "../form-field";
 import { makePrefixer, useId } from "../utils";
 import { RadioGroupContext } from "./internal/RadioGroupContext";
 import { RadioButton } from "./RadioButton";
-import { RadioIconProps } from "./RadioIcon";
+import { RadioButtonIconProps } from "./RadioButtonIcon";
 
 import "./RadioButtonGroup.css";
 
@@ -23,7 +23,7 @@ const withBaseName = makePrefixer("uitkRadioButtonGroup");
 export interface RadioButtonGroupProps extends HTMLAttributes<HTMLDivElement> {
   className?: string;
   defaultValue?: string;
-  icon?: FC<RadioIconProps>;
+  icon?: FC<RadioButtonIconProps>;
   legend?: string;
   name?: string;
   onChange?: ChangeEventHandler<HTMLInputElement>;
