+++ packages/core/src/radio-button/index.ts
@@ -1,4 +1,4 @@
 export * from "./RadioButton";
 export * from "./RadioButtonGroup";
-export * from "./Radio";
-export * from "./RadioIcon";
+export * from "./RadioButtonBase";
+export * from "./RadioButtonIcon";
